feat: 공통적인 의존성 주입/해제 로직 관리(base)
