+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -2,15 +2,13 @@ package woowacourse.shopping.ui.cart
 
 import android.os.Bundle
 import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import com.zzang.di.DIContainer.clearActivityScopedInstances
-import com.zzang.di.DependencyInjector
 import com.zzang.di.annotation.Inject
+import com.zzang.di.base.DIActivity
 import com.zzang.di.provideViewModel
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
 
-class CartActivity : AppCompatActivity() {
+class CartActivity : DIActivity() {
     private val binding by lazy { ActivityCartBinding.inflate(layoutInflater) }
 
     private val viewModel: CartViewModel by provideViewModel()
@@ -22,7 +20,6 @@ class CartActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        DependencyInjector.injectDependencies(this, this)
         setupBinding()
         setupToolbar()
         setupView()
@@ -70,10 +67,5 @@ class CartActivity : AppCompatActivity() {
             if (!it) return@observe
             Toast.makeText(this, getString(R.string.cart_deleted), Toast.LENGTH_SHORT).show()
         }
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        clearActivityScopedInstances(this)
     }
 }
