+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -2,17 +2,15 @@ package woowacourse.shopping.ui.cart
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.zzang.di.DIContainer
-import com.zzang.di.DependencyInjector
 import com.zzang.di.annotation.Inject
 import com.zzang.di.annotation.QualifierType
+import com.zzang.di.base.DIViewModel
 import kotlinx.coroutines.launch
 import woowacourse.shopping.domain.CartRepository
 import woowacourse.shopping.model.CartProduct
 
-class CartViewModel : ViewModel() {
+class CartViewModel : DIViewModel() {
     @Inject(qualifier = QualifierType.DATABASE)
     lateinit var cartRepository: CartRepository
 
@@ -21,10 +19,6 @@ class CartViewModel : ViewModel() {
 
     private val _onCartProductDeleted: MutableLiveData<Boolean> = MutableLiveData(false)
     val onCartProductDeleted: LiveData<Boolean> get() = _onCartProductDeleted
-
-    init {
-        DependencyInjector.injectDependencies(this, this)
-    }
 
     fun getAllCartProducts() {
         viewModelScope.launch {
@@ -42,10 +36,5 @@ class CartViewModel : ViewModel() {
                 _onCartProductDeleted.value = false
             }
         }
-    }
-
-    override fun onCleared() {
-        super.onCleared()
-        DIContainer.clearViewModelScopedInstances(this)
     }
 }
